fix: ex04 mod2

+++ modulo2/ex04/main.c
@@ -2,8 +2,8 @@
 #include "asm.h"
 
 int op1=0, op2=0;
-extern int op3;
-extern int op4;
+extern long op3;
+extern long op4;
 
 int main(void) {
 printf("Valor op1:");
@@ -11,9 +11,9 @@ scanf("%d",&op1);
 printf("Valor op2:");
 scanf("%d",&op2);
 printf("Valor op3:");
-scanf("%d",&op3);
+scanf("%ld",&op3);
 printf("Valor op4:");
-scanf("%d",&op4);
+scanf("%ld",&op4);
 
 long n = sum_v3();
